add environment variables to migration bot functions v1 to v2

+++ commands/administration/lastNewUsersChart.js
@@ -23,6 +23,8 @@ module.exports = class HelpCommand extends Command {
     }
 
     async run(msg, args) {
+        if (process.env.ENABLED_ADMINISTRATION !== 'true') return;
+        
         utils.logMessageUserExecuteCommand(logName, msg);
 
         msg.delete();
